Move interfaces for tasks to separate file
